TensorFlow: fix bug in TensorShapeDim -- size is the wrong type.
Adds a test that failed before this change, but passes after.
TensorShapeIter is only used by TensorShape::begin() and end(),
which are only used by AppendShape, which is only used by queue_base.h
on shapes that are unlikely to be 64-bit large, which is why
we never triggered this.

Discovered by jeff
Change: 113105906

+++ tensorflow/core/framework/tensor_shape.h
@@ -143,7 +143,7 @@ class TensorShape {
 
 struct TensorShapeDim {
   explicit TensorShapeDim(int64 s) : size(s) {}
-  int size;
+  int64 size;
 };
 
 class TensorShapeIter {
